Refactoring, improve hole handling
